Fix: pizzas spawning only on right half of screen. Change number of pizza columns too.
Pizzas were spawning on only the right half of the screen for larger screens. Set initial style.left of all pizzas to zero, preventing this.

Changed the number of pizza columns to a hardcoded 8, from a variable based on screen width. This prevents any potential issues with the animation math in updatePostions, should the number of pizza columns be a multiple of 5.